Alzheimer's Disease Dataset

Health, lifestyle, and clinical data for 2,149 individuals used for studying Alzheimer's Disease. Variables include demographics, cognitive assessments, medical conditions, and symptoms.

data(alzheimers)

Format

A data frame with 2,149 observations on the following variables:

Age

Age in years (60 to 90).

Gender

Gender (0 = Male, 1 = Female).

Ethnicity

Ethnicity (0 = Caucasian, 1 = African American, 2 = Asian, 3 = Other).

EducationLevel

Education level (0 = None, 1 = High School, 2 = Bachelor's, 3 = Higher).

BMI

Body Mass Index (15 to 40).

Smoking

Smoking status (0 = No, 1 = Yes).

AlcoholConsumption

Weekly alcohol consumption in units (0 to 20).

PhysicalActivity

Weekly physical activity in hours (0 to 10).

DietQuality

Diet quality score (0 to 10).

SleepQuality

Sleep quality score (4 to 10).

FamilyHistoryAlzheimers

Family history of Alzheimer's (0 = No, 1 = Yes).

CardiovascularDisease

Cardiovascular disease (0 = No, 1 = Yes).

Diabetes

Diabetes (0 = No, 1 = Yes).

Depression

Depression (0 = No, 1 = Yes).

HeadInjury

History of head injury (0 = No, 1 = Yes).

Hypertension

Hypertension (0 = No, 1 = Yes).

SystolicBP

Systolic blood pressure (90 to 180 mmHg).

DiastolicBP

Diastolic blood pressure (60 to 120 mmHg).

CholesterolTotal

Total cholesterol (150 to 300 mg/dL).

CholesterolLDL

LDL cholesterol (50 to 200 mg/dL).

CholesterolHDL

HDL cholesterol (20 to 100 mg/dL).

CholesterolTriglycerides

Triglycerides (50 to 400 mg/dL).

MMSE

Mini-Mental State Examination score (0 to 30). Lower is worse.

FunctionalAssessment

Functional score (0 to 10). Lower is worse.

MemoryComplaints

Memory complaints (0 = No, 1 = Yes).

BehavioralProblems

Behavioral problems (0 = No, 1 = Yes).

ADL

Activities of Daily Living score (0 to 10). Lower is worse.

Confusion

Presence of confusion (0 = No, 1 = Yes).

Disorientation

Presence of disorientation (0 = No, 1 = Yes).

PersonalityChanges

Presence of personality changes (0 = No, 1 = Yes).

DifficultyCompletingTasks

Difficulty completing tasks (0 = No, 1 = Yes).

Forgetfulness

Forgetfulness (0 = No, 1 = Yes).

Diagnosis

Alzheimer's diagnosis (No, Yes).

Details

This dataset is suitable for modeling Alzheimer's risk, performing exploratory analysis, and evaluating statistical and machine learning algorithms. All individuals are uniquely identified and evaluated on a standardized set of clinical and behavioral measures.

Source

Rabie El Kharoua (2024). Alzheimer's Disease Dataset. Available from Kaggle at https://www.kaggle.com/datasets/rabieelkharoua/alzheimers-disease-dataset

Examples

# \donttest{
## load the data
data(alzheimers, package = "varPro")
o <- varpro(Diagnosis~.,alzheimers)
imp <- importance(o)
print(imp)
# }