Plots for Unsupervised Data Visualization

Plots for unsupervised data visualization

plot.clusterpro(x, xvar.names, shrink=TRUE,
   col=TRUE, col.names=NULL, sort=TRUE, cex=FALSE, breaks=10, ... )

Arguments

x

clusterpro object returned from a previous call to clusterpro.

xvar.names

Names (or integer indices) of the x-variables to plot. Defaults to all variables.

shrink

Logical. If TRUE, shrinks the release variable to zero.

col

Logical. If TRUE, colors the points in the plot.

col.names

Variable used to color the plots. Defaults to the release variable. Can also be an integer index.

sort

Logical. If TRUE, sorts plots by variable importance.

cex

Numeric value to scale point size.

breaks

Number of breaks used when mapping colors to points.

...

Additional arguments passed to plot.

Details

Generates a two-dimensional visualization using UMAP applied to the enhanced data corresponding to a release variable. This provides a low-dimensional representation of the clustered structure derived from the rule-based transformation of the original data.

Author

Hemant Ishwaran

References

McInnes L., Healy J. and Melville J. (2018). UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction. ArXiv e-prints.

See also