Obtain Strength Array and Other Values from a VarPro Object

Used to parse values from a VarPro object.

varpro.strength(object,
                m.target = NULL,
                max.rules.tree = 150,
                max.tree = 150,
                stat = c("importance", "complement", "oob"),
                membership = FALSE,
                seed = NULL,
                do.trace = FALSE)

Arguments

object

rfsrc object

m.target

Character value for multivariate families specifying the target outcome to be used. If left unspecified, the algorithm will choose a default target.

max.rules.tree

Maximum number of rules per tree.

max.tree

Maximum number of trees used for extracting rules.

stat

Determines which statistic is output. Either importance, complement mean, or oob mean.

membership

Return OOB and complementary OOB membership indices for each rule?

seed

Seed for repeatability.

do.trace

Used to obtain detailed trace.

Details

Not intended for end-user use and is primarily meant for internal use by the package.

Examples

  
# \donttest{
## ------------------------------------------------------------
## regression example: boston housing
## ------------------------------------------------------------

## load the data
data(BostonHousing, package = "mlbench")

o <- rfsrc(medv~., BostonHousing, ntree=100)

## call varpro.strength
varpro.strength(object = o, max.rules.tree = 10, max.tree = 15)

# }